Up until Path.js
