+++ frontend/src/components/Path.js
@@ -1,0 +1,8 @@
+const app_name = 'ucf-go';
+exports.buildPath = function buildPath(route) {
+    if (process.env.NODE_ENV === 'production') {
+        return 'https://' + app_name + '.herokuapp.com/' + route;
+    } else {
+        return 'http://localhost:5000/' + route;
+    }
+};
