+++ frontend/src/pages/CardPage.js
@@ -1,14 +1,14 @@
 import React from 'react';
 import PageTitle from '../components/PageTitle';
 import LoggedInName from '../components/LoggedInName';
-import CardUI from '../components/CardUI';
+// import CardUI from '../components/CardUI';
 const CardPage = () => {
-  return (
-    <div>
-      <PageTitle />
-      <LoggedInName />
-      <CardUI />
-    </div>
-  );
+    return (
+        <div>
+            <PageTitle />
+            <LoggedInName />
+            {/* <CardUI /> */}
+        </div>
+    );
 };
 export default CardPage;
